+++ spec/test_app/never_called.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 echo "I'm never executed by any script"