+++ spec/test_app/scripts/cd.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 : "${this:=that}"
 : "${that:=thing}" && touch /dev/fd/1
@@ -7,7 +7,7 @@ dir="$(cd "${BASH_SOURCE%/*}" || : ; pwd -P)"
 todir=''
 printf -v todir -- '%s' "$(find "$dir" -type d | head -n 1)"
 
-(cd ../.. || : ; cd "$HOME" || : )
+(cd ../.. || : ; cd "$OLDPWD" || :)
 
 
 cd ../.. || :