+++ spec/test_app/scripts/delete.sh
@@ -1,8 +1,8 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-cat > tmp.sh <<'BASH'
-#!/bin/bash
-rm -v $0
+cat > tmp.sh <<BASH
+#!${BASH:-/usr/bin/env bash}
+rm -v \$0
 BASH
 
 chmod +x tmp.sh