+++ spec/test_app/scripts/nested/simple.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # basic test program